# server/src/openapi.yaml
openapi: 3.0.3
info:
  title: OrgTree API
  description: |
    RESTful API for OrgTree - an organizational directory and visualization platform.

    ## Authentication
    Most endpoints require JWT authentication via Bearer token in the Authorization header.

    ## Rate Limiting
    - Authentication endpoints: 5 requests per 15 minutes per IP
    - Password reset: 10 requests per 15 minutes per IP
    - General API: No specific limits (recommended: implement client-side throttling)

    ## Real-time Updates
    OrgTree uses Socket.IO for real-time synchronization. Connect to the same server
    with your JWT token to receive live updates for organizations you're a member of.
  version: 1.0.0
  contact:
    name: OrgTree Support
  license:
    name: GPL 3.0
  servers:
    - url: http://localhost:3001/api
      description: Development server
    - url: https://orgtree-app.onrender.com/api
      description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from login endpoint
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          example: 'Error description'
...
    Department:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        parentId:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        description:
          type: string
          nullable: true
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Person:
      type: object
      properties:
        id:
          type: string
          format: uuid
        departmentId:
          type: string
          format: uuid
        name:
          type: string
        title:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          nullable: true
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Invitation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [viewer, editor, admin]
        status:
          type: string
          enum: [pending, accepted, expired, cancelled]
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        organizationName:
          type: string
        invitedByName:
          type: string
    AuditLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        actorId:
          type: string
          format: uuid
        actorName:
          type: string
        actorEmail:
          type: string
        actionType:
          type: string
          enum: [created, updated, deleted, settings_changed]
        entityType:
          type: string
          enum: [department, person, member, org, invitation]
        entityId:
          type: string
        entityName:
          type: string
        details:
          type: object
        createdAt:
          type: string
          format: date-time
...
    OwnershipTransfer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        organizationName:
          type: string
        fromUserId:
          type: string
          format: uuid
        fromUserName:
          type: string
        toUserId:
          type: string
          format: uuid
        toUserName:
          type: string
        toUserEmail:
          type: string
        status:
          type: string
          enum: [pending, accepted, rejected, cancelled, expired]
        reason:
          type: string
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: 'Access token required'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: 'Resource not found'
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
security:
  - bearerAuth: []
paths:
...
  /public/org/{shareToken}:
    parameters:
      - name: shareToken
        in: path
        required: true
        schema:
          type: string

    get:
      tags: [Sharing]
      summary: Get public organization
      description: Get organization data via public share token. No authentication required.
      security: []
      responses:
        '200':
          description: Public organization data
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  createdAt:
                    type: string
                    format: date-time
                  departments:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/Department'
                        - type: object
                          properties:
                            people:
                              type: array
                              items:
                                $ref: '#/components/schemas/Person'
        '404':
          $ref: '#/components/responses/NotFound'
  /public/invitation/{token}:
    parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string

    get:
      tags: [Invitations]
      summary: Get public invitation
      description: Get invitation details by token. No authentication required.
      security: []
      responses:
        '200':
          description: Invitation details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invitation'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== USERS (SUPERUSER) ====================
...
  /ownership/transfers/pending:
    get:
      tags: [Ownership Transfers]
      summary: List pending transfers
      description: Get pending transfers where the current user is the recipient.
      responses:
        '200':
          description: List of pending transfers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OwnershipTransfer'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /ownership/transfers/{transferId}:
    parameters:
      - name: transferId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Ownership Transfers]
      summary: Get transfer details
      description: Get details of a specific transfer.
      responses:
        '200':
          description: Transfer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnershipTransfer'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /ownership/transfers/{transferId}/accept:
    parameters:
      - name: transferId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      tags: [Ownership Transfers]
      summary: Accept transfer
      description: Accept an ownership transfer. Current user must be the recipient.
      responses:
        '200':
          description: Transfer accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  transfer:
                    $ref: '#/components/schemas/OwnershipTransfer'
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /ownership/transfers/{transferId}/reject:
    parameters:
      - name: transferId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      tags: [Ownership Transfers]
      summary: Reject transfer
      description: Reject an ownership transfer.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Transfer rejected
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  transfer:
                    $ref: '#/components/schemas/OwnershipTransfer'
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /ownership/transfers/{transferId}/cancel:
    parameters:
      - name: transferId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      tags: [Ownership Transfers]
      summary: Cancel transfer
      description: Cancel a pending transfer. Requester must be the initiator (owner).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Transfer cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  transfer:
                    $ref: '#/components/schemas/OwnershipTransfer'
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /ownership/transfers/{transferId}/audit-log:
    parameters:
      - name: transferId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Ownership Transfers]
      summary: Get transfer audit log
      description: Get history of actions for this transfer.
      responses:
        '200':
          description: Audit log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditLog'
        '401':
          $ref: '#/components/responses/Unauthorized'